DBZ-8961 move in different packages
