+++ quarkus-debezium-parent/quarkus-debezium-engine-parent/deployment/src/main/java/io/quarkus/debezium/deployment/engine/EngineProcessor.java
@@ -62,12 +62,12 @@ import io.quarkus.debezium.deployment.dotnames.DebeziumDotNames;
 import io.quarkus.debezium.deployment.items.DebeziumConnectorBuildItem;
 import io.quarkus.debezium.deployment.items.DebeziumGeneratedInvokerBuildItem;
 import io.quarkus.debezium.deployment.items.DebeziumMediatorBuildItem;
-import io.quarkus.debezium.engine.CapturingHandlerProducer;
-import io.quarkus.debezium.engine.CapturingInvoker;
 import io.quarkus.debezium.engine.DebeziumRecorder;
-import io.quarkus.debezium.engine.DynamicCapturingInvokerSupplier;
 import io.quarkus.debezium.engine.FullyQualifiedTableNameResolver;
 import io.quarkus.debezium.engine.QualifiedTableNameResolverRecorder;
+import io.quarkus.debezium.engine.capture.CapturingHandlerProducer;
+import io.quarkus.debezium.engine.capture.CapturingInvoker;
+import io.quarkus.debezium.engine.capture.DynamicCapturingInvokerSupplier;
 import io.quarkus.deployment.GeneratedClassGizmoAdaptor;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
