+++ quarkus-debezium-parent/quarkus-debezium-engine-parent/runtime/src/main/java/io/quarkus/debezium/engine/capture/CapturingHandlerProducer.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.debezium.engine;
+package io.quarkus.debezium.engine.capture;
 
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -13,6 +13,8 @@ import jakarta.enterprise.inject.Instance;
 import jakarta.enterprise.inject.Produces;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
+
+import io.quarkus.debezium.engine.FullyQualifiedTableNameResolver;
 
 public class CapturingHandlerProducer {
 
@@ -24,7 +26,7 @@ public class CapturingHandlerProducer {
 
     @Produces
     @Singleton
-    public DefaultCapturingHandler produce() {
+    public DebeziumCapturingHandler produce() {
         CapturingInvokerRegistry capturingInvokerRegistry = new CapturingInvokerRegistry(invokers
                 .stream()
                 .collect(Collectors.toMap(CapturingInvoker::getFullyQualifiedTableName, Function.identity())));
