+++ quarkus-debezium-parent/quarkus-debezium-engine-parent/runtime/src/main/java/io/quarkus/debezium/engine/capture/CapturingInvokerRegistry.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.debezium.engine;
+package io.quarkus.debezium.engine.capture;
 
 import static io.debezium.runtime.Capturing.EVERYTHING_QUALIFIER;
 
@@ -16,29 +16,29 @@ import org.slf4j.LoggerFactory;
 
 import io.debezium.engine.RecordChangeEvent;
 
-public class CapturingInvokerRegistry {
+class CapturingInvokerRegistry {
     private static final Logger logger = LoggerFactory.getLogger(CapturingInvokerRegistry.class);
 
     private final Map<String, CapturingInvoker> invokers;
 
-    public CapturingInvokerRegistry(Map<String, CapturingInvoker> invokers) {
+    CapturingInvokerRegistry(Map<String, CapturingInvoker> invokers) {
         this.invokers = invokers;
     }
 
-    public CapturingInvoker get(String qualifier) {
+    CapturingInvoker get(String qualifier) {
         return invokers.getOrDefault(qualifier, invokers.getOrDefault(EVERYTHING_QUALIFIER, noOpInvoker(qualifier)));
     }
 
-    private CapturingInvoker noOpInvoker(String table) {
+    private CapturingInvoker noOpInvoker(String fullyQualifiedTableName) {
         return new CapturingInvoker() {
             @Override
             public void capture(RecordChangeEvent<SourceRecord> event) {
-                logger.warn("table {} not assigned to any handler", table);
+                logger.warn("table {} not assigned to any handler", fullyQualifiedTableName);
             }
 
             @Override
             public String getFullyQualifiedTableName() {
-                return table;
+                return fullyQualifiedTableName;
             }
         };
     }
