+++ quarkus-debezium-parent/quarkus-debezium-engine-parent/runtime/src/main/java/io/quarkus/debezium/engine/capture/DefaultCapturingHandler.java
@@ -4,17 +4,18 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.debezium.engine;
+package io.quarkus.debezium.engine.capture;
 
 import org.apache.kafka.connect.source.SourceRecord;
 
 import io.debezium.engine.RecordChangeEvent;
+import io.quarkus.debezium.engine.FullyQualifiedTableNameResolver;
 
-public class DefaultCapturingHandler implements DebeziumCapturingHandler {
+class DefaultCapturingHandler implements DebeziumCapturingHandler {
     private final CapturingInvokerRegistry registry;
     private final FullyQualifiedTableNameResolver resolver;
 
-    public DefaultCapturingHandler(CapturingInvokerRegistry registry, FullyQualifiedTableNameResolver resolver) {
+    DefaultCapturingHandler(CapturingInvokerRegistry registry, FullyQualifiedTableNameResolver resolver) {
         this.registry = registry;
         this.resolver = resolver;
     }
