+++ quarkus-debezium-parent/quarkus-debezium-engine-parent/runtime/src/main/java/io/quarkus/debezium/engine/capture/DynamicCapturingInvokerSupplier.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.debezium.engine;
+package io.quarkus.debezium.engine.capture;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.function.Supplier;
